useraccounts/requirements.txt
certifi==2019.3.9
dj-database-url>=0.5.0,<0.6.0
Django>=2.1.9,<2.2.0
django-bootstrap4>=0.0.8,<0.1.0
django-heroku>=0.3.1,<0.4.0
django-nocaptcha-recaptcha>=0.0.20,<0.1.0
entrypoints>=0.3,<0.4
flake8>=3.7.9,<3.8.0
gunicorn>=19.9.0,<20.0.0
mccabe>=0.6.1,<0.7.0
psycopg2>=2.8.2,<2.9.0
pycodestyle>=2.5.0,<2.6.0
pyflakes>=2.1.1,<2.2.0
pytz>=2019.1
whitenoise>=4.1.2,<4.2.0
wincertstore==0.2
